requirements.txt
google_generativeai
streamlit